Fixed MinGW compilation: now sphere doesn't crash anymore, it appears to work fine. Though, thread naming isn't done in this case, and this should be looked at (even if it's not extremely needed) in the future. Updated CMake generated vcxproj to set the correct folder for the debugger and updated compiler flags. Changed custom _32B and _64B compiler-set defines to _32BITS and _64BITS. Fixed the majority of the warnings reported by MinGW. Moved back some method definitions from CResourceBase.cpp to CResourceBase.h, in order to have them implicitly inlined. Moved also some methods from CException.cpp to .h for the same reason. Added clearer log messages for garbage collection, and logged start/finish of garbage collection and CWorld::Restock(), since they may require some time. Added a server mode for garbage collection.

+++ src/common/common.cpp
@@ -311,7 +311,7 @@ int CvtNUNICODEToSystem( tchar * pOut, int iSizeOutBytes, const NCHAR * pInp, in
 
 		// Flip all from network order.
 		wchar szBuffer[ 1024*8 ];
-		for ( ; iInp < CountOf(szBuffer) - 1 && iInp < iSizeInChars && pInp[iInp]; iInp++ )
+		for ( ; iInp < (int)CountOf(szBuffer) - 1 && iInp < iSizeInChars && pInp[iInp]; iInp++ )
 		{
 			szBuffer[iInp] = pInp[iInp];
 		}
